clean up import folder
