+++ Import/111/import/0005-config-gs.js
@@ -1,0 +1,25 @@
+
+
+/// sample, related to gs api server.
+
+var db_name = 'groceryDS';
+
+var lbTables = [
+  // 'User',
+	'accessToken',
+	// 'AccessToken',
+	'ACL', 'RoleMapping',
+	'Role',
+
+	// custom tables
+	'user', 'userCredential', 'userIdentity',
+
+	// 'userGrocery', 'userFav',
+	'Ingredient', 'Grocery', 'Department',
+
+];
+
+module.exports = {
+  db_name: db_name
+  lbTables: lbTables
+}
