+++ Import/111/import/0005-config-rapi.js
@@ -1,0 +1,23 @@
+;
+
+
+var lbTables = [
+
+];
+
+
+/// sample, related to recipe api server.
+
+var db_name = 'recipeDS';
+
+var lbTables = [
+  'user',
+	'Ingredient',
+	'Grocery',
+  'Department',
+];
+
+module.exports = {
+  db_name: db_name
+  lbTables: lbTables
+}
